Add assets.prefix options for css and fonts
Although rails assets are _generally_ in /assets they can be
configured to be served from elsewhere. This breaks the demo.

The reason for 2 separate options is that I use this gem in non-rails
projects, where my fonts and css aren't necessarily in the same
place.

Since we allow users to specify the css and font asset directories
separately, it makes sense that they should also be able to
separate them for the purpose of serving up the demo.